+++ backend/src/routes/worker_routes.py
@@ -2,12 +2,15 @@ from dataclasses import asdict
 from typing import Dict, List
 
 import humps
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, Depends, HTTPException
 from pydantic import TypeAdapter
 
 from core.worker import Worker, WorkerProperty
 from routes.api_model import WorkerMessage, WorkerPropertyMessage
 from scripts.setup_database import worker_db, worker_dimension_db, worker_property_db
+from services.authentication.authn_services import authn_verify_session
+from services.authentication.authn_types import SessionContainerType
+from services.authorization.authz_services import permit_check
 from services.deletion_services.delete_worker import (
     add_back_worker_property_to_constraint_build,
 )
@@ -18,10 +21,20 @@ from services.deletion_services.delete_worker import (
 router = APIRouter()
 
 
-@router.post("/workers")
-def create_worker() -> WorkerMessage:
-    worker_created = worker_db.create_worker()
-    wd_bool = worker_dimension_db.get_worker_dimensions_by_entry_type("bool")
+@router.post("/workers/teams/{team_id}")
+async def create_worker(
+    team_id: str,
+    worker: WorkerMessage,
+    session: SessionContainerType = Depends(authn_verify_session()),
+) -> WorkerMessage:
+    if not await permit_check(session.get_user_id(), "create-worker", "team", team_id):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to create a worker",
+        )
+    w_data = api_msg_to_worker(worker)
+    worker_created = worker_db.create_worker(w_data)
+    wd_bool = worker_dimension_db.get_worker_dimensions_by_entry_type("bool", team_id)
     wp_bool = []
     for wd in wd_bool:
         wp_bool.append(
@@ -30,9 +43,16 @@ def create_worker() -> WorkerMessage:
     return worker_and_properties_to_api_msg(worker_created, wp_bool)
 
 
-@router.get("/workers")
-def get_workers() -> List[WorkerMessage]:
-    workers = worker_db.get_workers()
+@router.get("/workers/teams/{team_id}")
+async def get_workers(
+    team_id: str,
+    session: SessionContainerType = Depends(authn_verify_session()),
+) -> List[WorkerMessage]:
+    if not await permit_check(session.get_user_id(), "read-workers", "team", team_id):
+        raise HTTPException(
+            status_code=403, detail="You do not have permission to get workers"
+        )
+    workers = worker_db.get_workers(team_id)
     workers_properties = [
         worker_property_db.get_worker_properties_by_worker_id(worker.id)
         for worker in workers
@@ -43,25 +63,44 @@ def get_workers() -> List[WorkerMessage]:
     ]
 
 
-@router.put("/workers/{worker_id}")
-def update_worker(worker_id: str, worker: WorkerMessage) -> WorkerMessage:
+@router.put("/workers/{worker_id}/teams/{team_id}")
+async def update_worker(
+    team_id: str,
+    worker_id: str,
+    worker: WorkerMessage,
+    session: SessionContainerType = Depends(authn_verify_session()),
+) -> WorkerMessage:
+    if not await permit_check(session.get_user_id(), "update-worker", "team", team_id):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to update a worker",
+        )
     existing_worker = worker_db.get_worker_by_id(worker_id)
     if not existing_worker:
         raise HTTPException(status_code=404, detail="Worker does not exist")
-    worker_data = api_msg_to_worker(worker)
-    updated_worker = worker_db.update_worker(worker_data)
+    w_data = api_msg_to_worker(worker)
+    updated_worker = worker_db.update_worker(w_data)
     worker_properties = worker_property_db.get_worker_properties_by_worker_id(
         updated_worker.id
     )
     return worker_and_properties_to_api_msg(updated_worker, worker_properties)
 
 
-@router.put("/workers/{worker_id}/properties/{worker_dimension_id}")
-def update_worker_property(
+@router.put("/workers/{worker_id}/properties/{worker_dimension_id}/teams/{team_id}")
+async def update_worker_property(
+    team_id: str,
     worker_id: str,
     worker_dimension_id: str,
     worker_property: WorkerPropertyMessage,
+    session: SessionContainerType = Depends(authn_verify_session()),
 ) -> WorkerPropertyMessage:
+    if not await permit_check(
+        session.get_user_id(), "update-worker-property", "team", team_id
+    ):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to update a worker property",
+        )
     wp_data = api_msg_to_worker_property(worker_property)
     if wp_data.id == "":
         worker = worker_db.get_worker_by_id(worker_id)
@@ -77,8 +116,17 @@ def update_worker_property(
     return worker_property_to_api_msg(new_wp)
 
 
-@router.delete("/workers/{worker_id}")
-def delete_worker(worker_id: str) -> Dict:
+@router.delete("/workers/{worker_id}/teams/{team_id}")
+async def delete_worker(
+    worker_id: str,
+    team_id: str,
+    session: SessionContainerType = Depends(authn_verify_session()),
+) -> Dict:
+    if not await permit_check(session.get_user_id(), "delete-worker", "team", team_id):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to delete a worker",
+        )
     delete_worker_service(worker_id)
     return {"message": "Worker deleted"}
